Fix synchronous produce selective receive (#25)
Relates to https://github.com/jeffgrunewald/stargate/issues/22.
This improves the selective receive so that it doesn't mistake
a completely different msg for a produce error.
I kept it simple for now. If you look at how `:gen.call` is implemented,
you'll see that is a lot more complex involving process monitors and
timeouts.
I wanted to add a timeout now, but this adds an extra param,
making the `produce/2` a `produce/3` which clashes with the existing
`produce/3`, so I left that out for now.